+++ src/main/java/lain/mods/peacefulsurface/impl/forge/ForgeWorldObj.java
@@ -58,7 +58,7 @@ public class ForgeWorldObj implements IWorldObj {
     }
 
     @Override
-    public String getBiomeName(double x, double y, double z) {
+    public String getBiomeName(int x, int y, int z) {
         ServerLevel o;
         if ((o = w.get()) == null)
             return "UNKNOWN";
@@ -66,7 +66,7 @@ public class ForgeWorldObj implements IWorldObj {
     }
 
     @Override
-    public int getLightLevel(double x, double y, double z) {
+    public int getLightLevel(int x, int y, int z) {
         ServerLevel o;
         if ((o = w.get()) == null)
             return 0;
@@ -74,7 +74,7 @@ public class ForgeWorldObj implements IWorldObj {
     }
 
     @Override
-    public int getBlockLight(double x, double y, double z) {
+    public int getBlockLight(int x, int y, int z) {
         ServerLevel o;
         if ((o = w.get()) == null)
             return 0;
@@ -82,7 +82,7 @@ public class ForgeWorldObj implements IWorldObj {
     }
 
     @Override
-    public int getSkyLight(double x, double y, double z) {
+    public int getSkyLight(int x, int y, int z) {
         ServerLevel o;
         if ((o = w.get()) == null)
             return 0;
